Update common-dotnet-troubleshooting.rst
We ran into several issues working with a customer, where the presence of another vendor's APM agent in the environment prevented SignalFx .NET instrumentation from working, even though the other APM agent was not running.   Nothing was written to the logs, and it took us time to figure out why instrumentation wasn't working.