remove hard coded username from code example
The code example to setup the database uses a hard-coded user name. The config example to configure the integration shows the username can be set to any user.
This change removes the hard coded username from the database setup example, to indicate the user is free to choose the name of the user for the integration.