Update lambda-ec2-collector-gateway.rst
In my testing, I found that it's necessary to use specify the protocol with the OTEL_EXPORTER_OTLP_TRACES_ENDPOINT environment variable, as well as port 4318 instead of 4317, since the protocol is http/protobuf instead of grpc. 

I also found that it's necessary to set the SPLUNK_METRICS_ENDPOINT to point to the gateway collector, otherwise the metrics will flow directly to the signalfx ingest endpoint.